+++ src/hardcopy.c
@@ -1650,7 +1650,7 @@ prt_flush_buffer(void)
 	    prt_write_string("s\n");
 
 	ga_clear(&prt_ps_buffer);
-	ga_init2(&prt_ps_buffer, (int)sizeof(char), prt_bufsiz);
+	ga_init2(&prt_ps_buffer, sizeof(char), prt_bufsiz);
     }
 }
 
@@ -2649,7 +2649,7 @@ mch_print_init(
     prt_bufsiz = psettings->chars_per_line;
     if (prt_out_mbyte)
 	prt_bufsiz *= 2;
-    ga_init2(&prt_ps_buffer, (int)sizeof(char), prt_bufsiz);
+    ga_init2(&prt_ps_buffer, sizeof(char), prt_bufsiz);
 
     prt_page_num = 0;
